v0.5.0: -u homedir, /boot mount warning, more robust cmdline.txt handler
This is mostly to address bugs here:

  https://forums.adafruit.com/viewtopic.php?f=50&t=70415

...and should leave the helper able to function on a Kali system, or other
places where filesystem layout differs a little from stock Raspbian
installations.

This won't solve all the problems of Kali users, since Kali is using its
own kernel, and the Adafruit kernel package is required for the PiTFT at
the moment, but it's probably a step in the right direction.